services/hardware-service/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# SSH
paramiko==3.4.0

# Serial port
pyserial==3.5

# HTTP client (for internal calls and health checks)
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
